Add debugging output to transform.c
Introduced a debugging output to display the values of semi-major and semi-minor axes in the transform.c file. The debug information will help track any potential issues related to these parameters during the transformation process.

+++ c/src/models/earth/earth.c
@@ -92,7 +92,7 @@ static PyObject* earth_model_set_ellipsoid(PyObject* self, PyObject* args) {
 
     PyObject* capsule;
     EarthModel* model;
-    long double a, b;
+    double a, b;
 
     if(!PyArg_ParseTuple(args, "Odd", &capsule, &a, &b)) {
         PyErr_SetString(PyExc_TypeError, "Invalid arguments passed to earth_model_set_ellipsoid.");
